pyproject.toml
[tool.poetry]
name = "duolingo-memcode-exporter"
version = "0.1.0"
description = ""
authors = ["Daniel Eder <danieleder@outlook.at>"]
readme = "README.md"
packages = [{include = "duolingo_memcode_exporter"}]

[tool.poetry.dependencies]
python = "^3.11"
duolingo-api = {git = "https://github.com/daniel-eder/Duolingo.git"}
googletrans = ">=4.0.0rc1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
